fix(runtime): move UiUpdate into runtime module surface
