+++ src/runtime/context.rs
@@ -1,13 +1,11 @@
-use crate::app::UiUpdate;
 use crate::state::ConversationManager;
-use tokio::sync::mpsc;
 
 pub struct RuntimeContext<'a> {
     pub conversation: &'a mut ConversationManager,
 }
 
 impl<'a> RuntimeContext<'a> {
-    pub fn start_turn(&mut self, _input: String, _tx: mpsc::UnboundedSender<UiUpdate>) {
+    pub fn start_turn(&mut self, _input: String) {
         // wired in REF-04
     }
